chore: move regression tests into relation test suite
