+++ tests/relations/many-to-one.test.ts
@@ -66,3 +66,64 @@ it('scopes a many-to-one relation update to the targeted record', async () => {
     { title: 'Second', author: { id: 2 } },
   ])
 })
+
+it('updates a many-to-one relation to a foreign record already associated with another owner', async () => {
+  const languageSchema = z.object({ code: z.string() })
+  const userSchema = z.object({
+    name: z.string(),
+    language: languageSchema,
+  })
+
+  const languages = new Collection({ schema: languageSchema })
+  const users = new Collection({ schema: userSchema })
+
+  users.defineRelations(({ one }) => ({
+    language: one(languages, { unique: false }),
+  }))
+
+  const langPt = await languages.create({ code: 'pt' })
+  const langEn = await languages.create({ code: 'en' })
+
+  const userOne = await users.create({ name: 'John', language: langPt })
+  await users.create({ name: 'Kate', language: langEn })
+
+  await users.update(userOne, {
+    data(user) {
+      user.language = langEn
+    },
+  })
+
+  expect(users.all()).toEqual([
+    { name: 'John', language: { code: 'en' } },
+    { name: 'Kate', language: { code: 'en' } },
+  ])
+})
+
+it("scopes nested updates to the updated owner's foreign record", async () => {
+  const countrySchema = z.object({ code: z.string() })
+  const userSchema = z.object({
+    name: z.string(),
+    country: countrySchema,
+  })
+
+  const countries = new Collection({ schema: countrySchema })
+  const users = new Collection({ schema: userSchema })
+
+  users.defineRelations(({ one }) => ({
+    country: one(countries),
+  }))
+
+  const us = await countries.create({ code: 'us' })
+  const ca = await countries.create({ code: 'ca' })
+
+  const userOne = await users.create({ name: 'John', country: us })
+  await users.create({ name: 'Kate', country: ca })
+
+  await users.update(userOne, {
+    data(user) {
+      user.country.code = 'uk'
+    },
+  })
+
+  expect(countries.all()).toEqual([{ code: 'uk' }, { code: 'ca' }])
+})
